requirements.txt
pillow==9.1.1
tensorflow==2.9.1
numpy==1.22.4
matplotlib==3.5.2
aiogram==2.20